refactor: different embed methods for core and app
In practice, there are now two entrypoints, which can be efficiently
tree-shaken, but Vite doesn't readily support that in library mode.
